requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
psutil==6.1.1
sse-starlette==2.2.1
pyobjc-core==10.3.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.3.1; sys_platform == 'darwin'
